move localKeySort to OrderedRVIterator
